pyproject.toml
[tool.poetry]
name = "ocrworker"
version = "0.1.0"
description = "OCR Worker"
authors = ["Eugen Ciur <eugen@papermerge.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ocrmypdf = "^16.3.1"
ocrmypdf-papermerge = "^0.6.0"
celery = {extras = ["redis"], version = "^5.4.0"}
sqlalchemy = "^2.0.30"
typer = "^0.12.3"
pydantic = "^2.7.4"
pydantic-settings = "^2.3.2"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
